ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
boto3==1.35.73
pydantic==2.10.4
python-dotenv==1.0.1
